refactor(QYOG-98): 순환 의존성 제거
