+++ src/components/Design/Pagination/emotion.ts
@@ -5,7 +5,7 @@
  */
 
 import styled from "@emotion/styled";
-import { Button, Row } from "@/components";
+
 import { lightThemeColor } from "@/styles/theme";
 
 export const PaginationButton = styled.button`
