+++ src/components/UI/Quill/Quill.tsx
@@ -1,8 +1,8 @@
 import dynamic from "next/dynamic";
 import "react-quill/dist/quill.snow.css";
 
-import { Button } from "@/components";
 import { validator } from "@/utils";
+import { Button } from "@/components/Design";
 
 const QuillNoSSRWrapper = validator.isClient
   ? dynamic(import("react-quill"), {
