--- src/components/index.ts
@@ -1,12 +1,0 @@
-/*
- * Created on Mon Nov 13 2023
- *
- * Copyright (c) 2023 Your Company
- */
-
-export * from "./UI";
-export * from "./Layouts";
-export * from "./Utilities";
-export * from "./Provider";
-export * from "./Svg";
-export * from "./Design";
